+++ app/models/folio/file_placement.rb
@@ -64,6 +64,7 @@ end
 #  position       :integer
 #  created_at     :datetime         not null
 #  updated_at     :datetime         not null
+#  type           :string
 #
 # Indexes
 #